Bench: 7072296

+++ Halogen/src/Network.cpp
@@ -22,7 +22,7 @@ Network InitNetwork(std::string file)
 
     std::string line;
 
-    std::vector<std::vector<double>> weights;
+    std::vector<std::vector<float>> weights;
     std::vector<size_t> LayerNeurons;
     weights.push_back({});
 
@@ -43,7 +43,7 @@ Network InitNetwork(std::string file)
 
         if (token == "HiddenLayerNeurons")
         {
-            std::vector<double> layerWeights;
+            std::vector<float> layerWeights;
 
             iss >> token;
             size_t num = stoull(token);
@@ -67,7 +67,7 @@ Network InitNetwork(std::string file)
         else if (token == "OutputLayer")
         {
             LayerNeurons.push_back(1);  //always 1 output neuron
-            std::vector<double> layerWeights;
+            std::vector<float> layerWeights;
             getline(stream, line);
             std::istringstream lineStream(line);
             while (lineStream >> token)
@@ -87,7 +87,7 @@ Network InitNetwork()
     std::string line;
     std::istringstream stream(oss.str());
 
-    std::vector<std::vector<double>> weights;
+    std::vector<std::vector<float>> weights;
     std::vector<size_t> LayerNeurons;
     weights.push_back({});
 
@@ -108,7 +108,7 @@ Network InitNetwork()
 
         if (token == "HiddenLayerNeurons")
         {
-            std::vector<double> layerWeights;
+            std::vector<float> layerWeights;
 
             iss >> token;
             size_t num = stoull(token);
@@ -132,7 +132,7 @@ Network InitNetwork()
         else if (token == "OutputLayer")
         {
             LayerNeurons.push_back(1);  //always 1 output neuron
-            std::vector<double> layerWeights;
+            std::vector<float> layerWeights;
             getline(stream, line);
             std::istringstream lineStream(line);
             while (lineStream >> token)
@@ -146,37 +146,37 @@ Network InitNetwork()
     return Network(weights, LayerNeurons);
 }
 
-Neuron::Neuron(std::vector<double> Weight, double Bias) : grad(Weight.size() + 1, 0)
+Neuron::Neuron(std::vector<float> Weight, float Bias) : grad(Weight.size() + 1, 0)
 {
     weights = Weight;
     bias = Bias;
 }
 
-double Neuron::FeedForward(std::vector<double>& input) const
+float Neuron::FeedForward(std::vector<float>& input) const
 {
     assert(input.size() == weights.size());
 
-    double ret = bias;
+    float ret = bias;
 
     for (size_t i = 0; i < input.size(); i++)
     {
-        ret += std::max(0.0, input[i]) * weights[i];
+        ret += std::max(0.f, input[i]) * weights[i];
     }
 
     return ret;
 }
 
-void Neuron::Backpropogate(double delta_l, const std::vector<double>& prev_weights, double learnRate)
+void Neuron::Backpropogate(float delta_l, const std::vector<float>& prev_weights, float learnRate)
 {
     for (size_t weight = 0; weight < weights.size(); weight++)
     {
-        double new_grad = delta_l * std::max(0.0, prev_weights.at(weight)); //ReLU activation calculated here
+        float new_grad = delta_l * std::max(0.f, prev_weights.at(weight)); //ReLU activation calculated here
 
         grad.at(weight) += new_grad * new_grad;
         weights.at(weight) -= new_grad * learnRate *InvSqrt(grad.at(weight) + 10e-8);
     }
 
-    double new_grad = delta_l;
+    float new_grad = delta_l;
     grad.at(weights.size()) += new_grad * new_grad;
     bias -= new_grad * learnRate *InvSqrt(grad.at(weights.size()) + 10e-8);
 }
@@ -190,7 +190,7 @@ void Neuron::WriteToFile(std::ofstream& myfile)
     myfile << bias << "\n";
 }
 
-HiddenLayer::HiddenLayer(std::vector<double> inputs, size_t NeuronCount)
+HiddenLayer::HiddenLayer(std::vector<float> inputs, size_t NeuronCount)
 {
     assert(inputs.size() % NeuronCount == 0);
 
@@ -198,7 +198,7 @@ HiddenLayer::HiddenLayer(std::vector<double> inputs, size_t NeuronCount)
 
     for (size_t i = 0; i < NeuronCount; i++)
     {
-        neurons.push_back(Neuron(std::vector<double>(inputs.begin() + (WeightsPerNeuron * i), inputs.begin() + (WeightsPerNeuron * i) + WeightsPerNeuron - 1), inputs.at(WeightsPerNeuron * (1 + i) - 1)));
+        neurons.push_back(Neuron(std::vector<float>(inputs.begin() + (WeightsPerNeuron * i), inputs.begin() + (WeightsPerNeuron * i) + WeightsPerNeuron - 1), inputs.at(WeightsPerNeuron * (1 + i) - 1)));
     }
 
     for (size_t i = 0; i < WeightsPerNeuron - 1; i++)
@@ -209,10 +209,10 @@ HiddenLayer::HiddenLayer(std::vector<double> inputs, size_t NeuronCount)
         }
     }
 
-    zeta = std::vector<double>(NeuronCount, 0);
+    zeta = std::vector<float>(NeuronCount, 0);
 }
 
-std::vector<double> HiddenLayer::FeedForward(std::vector<double>& input)
+std::vector<float> HiddenLayer::FeedForward(std::vector<float>& input)
 {
     for (size_t i = 0; i < neurons.size(); i++)
     {
@@ -222,7 +222,7 @@ std::vector<double> HiddenLayer::FeedForward(std::vector<double>& input)
     return zeta;
 }
 
-void HiddenLayer::Backpropogate(const std::vector<double>& delta_l, const std::vector<double>& prev_weights, double learnRate)
+void HiddenLayer::Backpropogate(const std::vector<float>& delta_l, const std::vector<float>& prev_weights, float learnRate)
 {
     assert(delta_l.size() == neurons.size());
 
@@ -242,19 +242,19 @@ void HiddenLayer::WriteToFile(std::ofstream& myfile)
     }
 }
 
-void HiddenLayer::activation(const std::vector<double>& in, std::vector<double>& out)
+void HiddenLayer::activation(const std::vector<float>& in, std::vector<float>& out)
 {
     assert(in.size() == out.size());
 
     for (size_t i = 0; i < in.size(); i++)
     {
-        out[i] = std::max(0.0, in[i]);
+        out[i] = std::max(0.f, in[i]);
     }
 }
 
-std::vector<double> HiddenLayer::activationPrime(std::vector<double> x)
+std::vector<float> HiddenLayer::activationPrime(std::vector<float> x)
 {
-    std::vector<double> ret;
+    std::vector<float> ret;
     ret.reserve(x.size());
 
     for (size_t i = 0; i < x.size(); i++)
@@ -265,7 +265,7 @@ std::vector<double> HiddenLayer::activationPrime(std::vector<double> x)
     return ret;
 }
 
-void HiddenLayer::ApplyDelta(std::vector<deltaPoint>& deltaVec, double forward)
+void HiddenLayer::ApplyDelta(std::vector<deltaPoint>& deltaVec, float forward)
 {
     assert(deltaVec.size() == INPUT_NEURONS);
     size_t neuronCount = zeta.size();
@@ -273,7 +273,7 @@ void HiddenLayer::ApplyDelta(std::vector<deltaPoint>& deltaVec, double forward)
 
     for (size_t index = 0; index < deltaCount; index++)
     {
-        double deltaValue = deltaVec[index].delta * forward;
+        float deltaValue = deltaVec[index].delta * forward;
         size_t weightTransposeIndex = deltaVec[index].index * neuronCount;
 
         for (size_t neuron = 0; neuron < neuronCount; neuron++)
@@ -283,7 +283,7 @@ void HiddenLayer::ApplyDelta(std::vector<deltaPoint>& deltaVec, double forward)
     }
 }
 
-Network::Network(std::vector<std::vector<double>> inputs, std::vector<size_t> NeuronCount) : outputNeuron(std::vector<double>(inputs.back().begin(), inputs.back().end() - 1), inputs.back().back())
+Network::Network(std::vector<std::vector<float>> inputs, std::vector<size_t> NeuronCount) : outputNeuron(std::vector<float>(inputs.back().begin(), inputs.back().end() - 1), inputs.back().back())
 {
     assert(inputs.size() == NeuronCount.size());
 
@@ -297,7 +297,7 @@ Network::Network(std::vector<std::vector<double>> inputs, std::vector<size_t> Ne
     }
 }
 
-double Network::FeedForward(std::vector<double> inputs)
+float Network::FeedForward(std::vector<float> inputs)
 {
     assert(inputs.size() == inputNeurons);
 
@@ -312,21 +312,21 @@ double Network::FeedForward(std::vector<double> inputs)
     return zeta;
 }
 
-double Network::Backpropagate(trainingPoint data, double learnRate)
+float Network::Backpropagate(trainingPoint data, float learnRate)
 {
-    std::vector<double> inputParams(data.inputs.begin(), data.inputs.end());
+    std::vector<float> inputParams(data.inputs.begin(), data.inputs.end());
 
     FeedForward(inputParams);
     //return 0.5 * (alpha - data.result) * (alpha - data.result);   //if you just want to calculate error without training then do this
 
     //we choose a vector for this because delta_l will have a value for each neuron in the layer later
-    std::vector<double> delta_l = { 0.0087 * (alpha - data.result) * (alpha) * (1 - alpha) }; //0.0087 chosen to mymic the previous evaluation function
+    std::vector<float> delta_l = { 0.0087f * (alpha - data.result) * (alpha) * (1 - alpha) }; //0.0087 chosen to mymic the previous evaluation function
     outputNeuron.Backpropogate(delta_l[0], hiddenLayers.back().zeta, learnRate);
 
     for (int layer = hiddenLayers.size() - 1; layer >= 0; layer--)
     {
-        std::vector<double> tmp = hiddenLayers.at(layer).activationPrime(hiddenLayers.at(layer).zeta);  //1 if this neuron is on and 0 if its not
-        std::vector<double> tmp2;
+        std::vector<float> tmp = hiddenLayers.at(layer).activationPrime(hiddenLayers.at(layer).zeta);  //1 if this neuron is on and 0 if its not
+        std::vector<float> tmp2;
 
         for (int neuron = 0; neuron < hiddenLayers.at(layer).neurons.size(); neuron++)
         {
@@ -336,7 +336,7 @@ double Network::Backpropagate(trainingPoint data, double learnRate)
             }
             else
             {
-                double val = 0;
+                float val = 0;
 
                 for (int nextLayerNeuron = 0; nextLayerNeuron < hiddenLayers.at(layer + 1).neurons.size(); nextLayerNeuron++)
                 {
@@ -348,7 +348,7 @@ double Network::Backpropagate(trainingPoint data, double learnRate)
         }
 
         delta_l.clear();
-        std::transform(tmp.begin(), tmp.end(), tmp2.begin(), std::back_inserter(delta_l), std::multiplies<double>()); //element-wise multiplication of delta_l = tmp * tmp2;
+        std::transform(tmp.begin(), tmp.end(), tmp2.begin(), std::back_inserter(delta_l), std::multiplies<float>()); //element-wise multiplication of delta_l = tmp * tmp2;
 
         if (layer > 0)
             hiddenLayers.at(layer).Backpropogate(delta_l, hiddenLayers.at(layer - 1).zeta, learnRate); //get the previous layers activation
@@ -413,7 +413,7 @@ void Network::Learn()
         auto rng = std::default_random_engine{};
         std::shuffle(std::begin(data), std::end(data), rng);
 
-        double error = 0;
+        float error = 0;
 
         for (size_t point = 0; point < data.size() / 100; point++)
         {
@@ -443,9 +443,9 @@ void Network::ApplyInverseDelta(std::vector<deltaPoint>& delta)
     hiddenLayers[0].ApplyDelta(delta, -1);
 }
 
-double Network::QuickEval()
+float Network::QuickEval()
 {
-    std::vector<double>& inputs = hiddenLayers.at(0).zeta;
+    std::vector<float>& inputs = hiddenLayers.at(0).zeta;
 
     for (size_t i = 1; i < hiddenLayers.size(); i++)    //skip first layer
     {
@@ -458,7 +458,7 @@ double Network::QuickEval()
     return zeta;
 }
 
-trainingPoint::trainingPoint(std::array<bool, INPUT_NEURONS> input, double gameResult)
+trainingPoint::trainingPoint(std::array<bool, INPUT_NEURONS> input, float gameResult)
 {
     inputs = input;
     result = gameResult;
@@ -469,7 +469,7 @@ Network CreateRandom(std::vector<size_t> NeuronCount)
     std::random_device rd;
     std::mt19937 e2(rd());
 
-    std::vector<std::vector<double>> inputs;
+    std::vector<std::vector<float>> inputs;
     inputs.push_back({});
 
     size_t prevLayerNeurons = NeuronCount[0];
@@ -477,9 +477,9 @@ Network CreateRandom(std::vector<size_t> NeuronCount)
     for (size_t layer = 1; layer < NeuronCount.size() - 1; layer++)
     {
         //see https://arxiv.org/pdf/1502.01852v1.pdf eq (10) for theoretical significance of w ~ N(0, sqrt(2/n))
-        std::normal_distribution<> dist(0.0, sqrt(2.0 / static_cast<double>(NeuronCount[layer])));
+        std::normal_distribution<> dist(0.0, sqrt(2.0 / static_cast<float>(NeuronCount[layer])));
 
-        std::vector<double> input;
+        std::vector<float> input;
         for (size_t i = 0; i < (prevLayerNeurons + 1) * NeuronCount[layer]; i++)
         {
             if ((i + 1) % (prevLayerNeurons + 1) == 0)
@@ -491,10 +491,10 @@ Network CreateRandom(std::vector<size_t> NeuronCount)
         prevLayerNeurons = NeuronCount[layer];
     }
 
-    std::normal_distribution<> dist(0.0, sqrt(2.0 / static_cast<double>(NeuronCount.back())));
+    std::normal_distribution<> dist(0.0, sqrt(2.0 / static_cast<float>(NeuronCount.back())));
 
     //connections from last hidden to output
-    std::vector<double> input;
+    std::vector<float> input;
     for (size_t i = 0; i < (prevLayerNeurons + 1) * NeuronCount.back(); i++)
     {
         if ((i + 1) % (prevLayerNeurons + 1) == 0)
@@ -509,7 +509,7 @@ Network CreateRandom(std::vector<size_t> NeuronCount)
 
 void Network::AddExtraNullLayer(size_t neurons)
 {
-    std::vector<double> weights;
+    std::vector<float> weights;
 
     for (int j = 0; j < neurons; j++)
     {
