+++ Halogen/src/Position.h
@@ -56,7 +56,7 @@ public:
 
 	Network net;
 
-	double GetEvaluation();
+	float GetEvaluation();
 
 private:
 	uint64_t NodeCount;
@@ -67,7 +67,7 @@ private:
 	uint64_t GenerateZobristKey() const;
 	uint64_t IncrementZobristKey(Move move);	
 
-	std::vector<double> GetInputLayer();
+	std::vector<float> GetInputLayer();
 	std::vector<deltaPoint> CalculateMoveDelta(Move move);				//A vector which calculates the CHANGE in each input paramiter
 
 	size_t modifier(size_t index);						//no inputs for pawns on front or back rank for neural net: we need to modify zobrist-like indexes
